+++ Hook/ConfigurationHook.php
@@ -21,44 +21,61 @@
 /*                                                                                */
 /**********************************************************************************/
 
+namespace DatabasesManager\Hook;
+
+use Thelia\Core\Event\Hook\HookRenderEvent;
+use Thelia\Core\Hook\BaseHook;
+
 /**
- * Locale for english - United States
+ * Class ConfigurationHook
  *
  * @author Jérôme Billiras <jerome DOT billiras PLUS github AT gmail DOT com>
  */
-return [
-    'SHARED_VIEW_TITLE' => 'Databases shared configuration.',
-    'ENV_VIEW_TITLE' => 'Databases configuration for current environment.',
+class ConfigurationHook extends BaseHook
+{
+    /**
+     * @var string Current environment
+     */
+    protected $environment;
 
-    'NO_CONFIG' => 'No configuration',
+    /**
+     * Class constructor
+     *
+     * @param string $environment Current environment
+     */
+    public function __construct($environment)
+    {
+        $this->environment = $environment;
+    }
 
-    'TABLE_HEAD_LABEL' => 'Label',
-    'TABLE_HEAD_HOST' => 'Host',
-    'TABLE_HEAD_USER' => 'User',
-    'TABLE_HEAD_PASS' => 'Password',
-    'TABLE_HEAD_DB_NAME' => 'Database name',
-    'TABLE_HEAD_ACTION' => 'Action',
+    /**
+     * Add module configuration content
+     *
+     * @param HookRenderEvent $hookRenderEvent
+     */
+    public function onModuleConfiguration(HookRenderEvent $hookRenderEvent)
+    {
+        $hookRenderEvent->add(
+            $this->render(
+                'module_configuration.html',
+                [
+                    'kernelEnvironment' => $this->environment
+                ]
+            )
+        );
+    }
 
-    'PWDS_ACTION_HIDE' => 'Hide passwords',
-    'PWD_ACTION_SHOW' => 'Show password',
-    'PWD_ACTION_HIDE' => 'Hide password',
-
-    'ACTION_EDIT' => 'Edit database',
-    'ACTION_DELETE' => 'Delete database',
-
-    'FORM_LABEL_LABEL' => 'Label',
-    'FORM_LABEL_HOST' => 'Host',
-    'FORM_LABEL_USER' => 'User',
-    'FORM_LABEL_PASS' => 'Password',
-    'FORM_LABEL_DB_NAME' => 'Database name',
-
-    'MODAL_TITLE_ADD' => 'Add a database configuration',
-    'CONFIG_EXISTS' => 'Database configuration with label "%label" already exists. Use editing or change the current label.',
-
-    'MODAL_TITLE_EDIT' => 'Edit a database configuration',
-
-    'MODAL_TITLE_DELETE' => 'Delete a database configuration',
-    'MODAL_DELETE_CONFIRM' => 'Do you really want to delete the database configuration &quot;%label&quot;&nbsp;?',
-    'CONFIG_DOES_NOT_EXIST' => 'Database configuration with label "%label" doesn\'t exist.',
-    'CONFIG_DELETED' => 'Database configuration with label "%label" was deleted.',
-];
+    /**
+     * Add module configuration content script
+     *
+     * @param HookRenderEvent $hookRenderEvent
+     */
+    public function onModuleConfigJs(HookRenderEvent $hookRenderEvent)
+    {
+        $hookRenderEvent->add(
+            $this->render(
+                'module-config-js.html'
+            )
+        );
+    }
+}